增加插件
