+++ HongMouer.Common/HongMouer.Common.csproj
@@ -15,4 +15,18 @@
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
   </PropertyGroup>
 
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
+    <OutputPath>C:\HongMouer\HongMouer.EHR.Models\Package\</OutputPath>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="HongMouer.EntityRelationalCore" Version="2020.8.19.2033" />
+    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
+    <PackageReference Include="System.Text.Json" Version="4.7.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\HongMouer.EHR.Models\HongMouer.EHR.Models.csproj" />
+  </ItemGroup>
+
 </Project>
