+++ HongMouer.DataAnnotations/HongMouer.DataAnnotations.csproj
@@ -15,4 +15,8 @@
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
   </PropertyGroup>
 
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
+    <OutputPath>C:\HongMouer\HongMouer.EHR.Models\Package\</OutputPath>
+  </PropertyGroup>
+
 </Project>
