+++ HongMouer.EHR.Core.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30309.148
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HongMouer.EHR.Models", "HongMouer.EHR.Models\HongMouer.EHR.Models.csproj", "{539A5C18-7EDB-4C1D-A126-F32822F7B074}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HongMouer.Common", "HongMouer.Common\HongMouer.Common.csproj", "{9ED9D003-5D53-4A06-91E0-580D5B2DD96C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HongMouer.DataAnnotations", "HongMouer.DataAnnotations\HongMouer.DataAnnotations.csproj", "{1CDA3B94-322D-4682-8AFD-AC7F4903836B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{539A5C18-7EDB-4C1D-A126-F32822F7B074}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{539A5C18-7EDB-4C1D-A126-F32822F7B074}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{539A5C18-7EDB-4C1D-A126-F32822F7B074}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{539A5C18-7EDB-4C1D-A126-F32822F7B074}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9ED9D003-5D53-4A06-91E0-580D5B2DD96C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9ED9D003-5D53-4A06-91E0-580D5B2DD96C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9ED9D003-5D53-4A06-91E0-580D5B2DD96C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9ED9D003-5D53-4A06-91E0-580D5B2DD96C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1CDA3B94-322D-4682-8AFD-AC7F4903836B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1CDA3B94-322D-4682-8AFD-AC7F4903836B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1CDA3B94-322D-4682-8AFD-AC7F4903836B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1CDA3B94-322D-4682-8AFD-AC7F4903836B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6391DB34-3643-4AE3-BD02-E390173E066D}
+	EndGlobalSection
+EndGlobal
