+++ Package/PublishNuget.bat
@@ -1,0 +1,1 @@
+nuget.exe push *.nupkg 8c4f2b59c5a544a0b723dad2a37b21f8 -Source http://114.67.103.214:8080/nuget
